add a function to handle range defaults
